+++ src/app/(logged-in)/my/general/organization-image-picker.tsx
@@ -4,7 +4,7 @@ import {useOverlayTrigger} from 'react-aria';
 import {type Organization} from '@prisma/client';
 import AddPhotoAlternate from '@material-design-icons/svg/round/add_photo_alternate.svg';
 import Save from '@material-design-icons/svg/round/save.svg';
-import Button from '@/components/button.tsx';
+import Button from '@/components/button/button.tsx';
 import Modal from '@/components/modal/modal.tsx';
 import ImageButton from '@/components/image-button.tsx';
 import Dialog from '@/components/dialog.tsx';
