+++ src/app/(logged-in)/new-organization/page.tsx
@@ -3,7 +3,7 @@ import OrganizationCreationForm from '@/components/organization-creation-form.ts
 import createOrganizationAction from '@/lib/actions/create-organization-action.ts';
 import {getUserFromSession} from '@/lib/models/user.ts';
 import TopBar from '@/components/top-bar.tsx';
-import LinkButton from '@/components/link-button.tsx';
+import LinkButton from '@/components/button/link-button.tsx';
 import Footer from '@/components/footer.tsx';
 
 export default async function NewOrganizationPage() {
