+++ src/app/(logged-in)/onboarding/account-button.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Link from 'next/link';
 import Person from '@material-design-icons/svg/round/person.svg';
 import ArrowDropDown from '@material-design-icons/svg/round/arrow_drop_down.svg';
-import Button from '@/components/button.tsx';
+import Button from '@/components/button/button.tsx';
 import PopoverButtonTrigger from '@/components/popover-button-trigger.tsx';
 
 export default function AccountButton() {
