+++ src/components/modal/modal-trigger.tsx
@@ -3,8 +3,8 @@ import React, {type ReactElement, type ReactNode} from 'react';
 import {type OverlayTriggerProps, useOverlayTriggerState} from 'react-stately';
 import {useOverlayTrigger} from 'react-aria';
 import Modal, {type ModalProps} from '@/components/modal/modal.tsx';
-import Button from '@/components/button.tsx';
-import {type ButtonVariantProps} from '@/components/variants/button.tsx';
+import Button from '@/components/button/button.tsx';
+import {type ButtonVariantProps} from '@/components/button/button-variants.tsx';
 import {modalContext} from '@/components/modal/modal-context.ts';
 
 export type ModalTriggerProps = {
