+++ src/components/number-field.tsx
@@ -5,7 +5,7 @@ import {useObjectRef} from '@react-aria/utils';
 import {twJoin} from 'tailwind-merge';
 import ArrowDropDown from '@material-design-icons/svg/round/arrow_drop_down.svg';
 import ArrowDropUp from '@material-design-icons/svg/round/arrow_drop_up.svg';
-import Button from '@/components/button.tsx';
+import Button from '@/components/button/button.tsx';
 import {cx} from '@/lib/cva.ts';
 
 export type NumberFieldProps = {
@@ -47,9 +47,9 @@ export const NumberField = React.forwardRef((
 					isRequired && 'after:content-["*"] after:ml-0.5',
 				)}>{label}</label>
 
-			<div {...groupProps} className='flex items-center justify-right ps-1 gap-1 rounded border border-stone-700 group-focus-within:border-stone-50 w-full group-data-[disabled=true]:border-stone-800'>
+			<div {...groupProps} className='flex items-center justify-right ps-2 gap-2 rounded border border-stone-700 group-focus-within:border-stone-50 w-full group-data-[disabled=true]:border-stone-800'>
 				{icon}
-				<input {...inputProps} ref={inputRef} name={name} className='bg-transparent text-stone-200 py-1 outline-none grow min-w-0 disabled:text-stone-600 disabled:cursor-not-allowed'/>
+				<input {...inputProps} ref={inputRef} name={name} className='bg-transparent text-stone-200 py-2 outline-none grow min-w-0 disabled:text-stone-600 disabled:cursor-not-allowed'/>
 				<div className='flex-none flex-col justify-around items-center basis-4 fill-stone-400'>
 					<Button {...incrementButtonProps} variant='text' size='xs'>
 						<ArrowDropUp viewBox='0 0 24 24' className='w-4 h-4'/>
