+++ src/components/search-field.tsx
@@ -3,7 +3,7 @@ import {type AriaSearchFieldProps, useSearchField} from 'react-aria';
 import {type SearchFieldProps as SearchFieldStateProps, useSearchFieldState} from 'react-stately';
 import Search from '@material-design-icons/svg/round/search.svg';
 import Close from '@material-design-icons/svg/round/close.svg';
-import Button from '@/components/button.tsx';
+import Button from '@/components/button/button.tsx';
 import {cx} from '@/lib/cva.ts';
 
 export type SearchFieldProps = {
