+++ src/components/submit-button.tsx
@@ -1,7 +1,7 @@
 'use client';
 import React, {type ReactNode} from 'react';
 import {useFormStatus} from 'react-dom';
-import Button, {type ButtonProps} from '@/components/button.tsx';
+import Button, {type ButtonProps} from '@/components/button/button.tsx';
 import LoadingSpinner from '@/components/loading-spinner.tsx';
 
 export type SubmitButtonProps = {
@@ -17,29 +17,19 @@ export default function SubmitButton(props: SubmitButtonProps) {
 		<Button {...props} isDisabled={isDisabled ?? pending} type='submit'>
 			{
 				iconPlacement === 'left' && (
-					<>
-						{
-							pending
-								? (
-									<LoadingSpinner className='m-1'/>
-								) : icon
-						}
-						<span className='w-1'/>
-					</>
+					pending
+						? (
+							<LoadingSpinner className='m-1'/>
+						) : icon
 				)
 			}
 			{children}
 			{
 				iconPlacement === 'right' && (
-					<>
-						<span className='w-1'/>
-						{
-							pending
-								? (
-									<LoadingSpinner className='m-1'/>
-								) : icon
-						}
-					</>
+					pending
+						? (
+							<LoadingSpinner className='m-1'/>
+						) : icon
 				)
 			}
 		</Button>
